# docs/runtime_api_openapi_snapshot.json
{
  "components": {
    "schemas": {
      "AssetAvailabilityRequest": {
        "properties": {
          "assets": {
            "items": {
              "type": "string"
            },
            "title": "Assets",
            "type": "array"
          }
        },
        "title": "AssetAvailabilityRequest",
        "type": "object"
      },
      "AssetDescriptionPatch": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          }
        },
        "title": "AssetDescriptionPatch",
        "type": "object"
      },
...
      "ColumnDescriptionPatch": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          }
        },
        "title": "ColumnDescriptionPatch",
        "type": "object"
      },
      "ColumnMetadataPatch": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          }
        },
        "title": "ColumnMetadataPatch",
        "type": "object"
      },
      "ColumnTagsPatch": {
        "properties": {
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tags",
            "type": "object"
          }
        },
        "title": "ColumnTagsPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Governance Atlas Runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assets/availability": {
      "post": {
        "operationId": "api_asset_availability_api_assets_availability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetAvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Asset Availability",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/assets/{asset_fqn}": {
      "get": {
        "operationId": "api_asset_detail_api_assets__asset_fqn__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sections",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Sections",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Asset Detail",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/assets/{asset_fqn}/columns/{column_name}/description": {
      "patch": {
        "operationId": "api_patch_column_description_api_assets__asset_fqn__columns__column_name__description_patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "column_name",
            "required": true,
            "schema": {
              "title": "Column Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnDescriptionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Patch Column Description",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_fqn}/columns/{column_name}/metadata": {
      "patch": {
        "operationId": "api_patch_column_metadata_api_assets__asset_fqn__columns__column_name__metadata_patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "column_name",
            "required": true,
            "schema": {
              "title": "Column Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnMetadataPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Patch Column Metadata",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_fqn}/columns/{column_name}/tags": {
      "patch": {
        "operationId": "api_patch_column_tags_api_assets__asset_fqn__columns__column_name__tags_patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "column_name",
            "required": true,
            "schema": {
              "title": "Column Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnTagsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Patch Column Tags",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/assets/{asset_fqn}/description": {
      "patch": {
        "operationId": "api_patch_asset_description_api_assets__asset_fqn__description_patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDescriptionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Patch Asset Description",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_fqn}/metadata": {
      "patch": {
        "operationId": "api_patch_asset_metadata_api_assets__asset_fqn__metadata_patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_fqn",
            "required": true,
            "schema": {
              "title": "Asset Fqn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetMetadataPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Patch Asset Metadata",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}